requirements.txt
docker==6.*
PyYAML==6.*
